rename ui_test::comments → parser
